Py3 (#40)
* Ignore backup files (*.bak)

* Use python 3 for default Dockerfile

* Add python 2 -based Dockerfile

* Add python 2 -based service in docker-compose.yml

* Add future lib

* Futurized stage1

* Futurized stage2

* Workaround to compute metadata + file hashes

In python 2 unicode strings are represented with an explicit unicode
literal ('u'), whereas in python 3, although supported, it is dropped by
default. Why is this a problem? In python 2, the hash of (metadata + file
hash) would be computed using an input such as:
"[u'artist name', u'piece title', 'file hash']". In python 3 the same
data would be: "['artist name', 'piece title', 'file hash']". The fact
that the prefix 'u' is not there results in a different hash. In other
words, the inputs are different because the reprsentations are
different, which yields different outputs.

* Add python 3.5 to travis

* Add source encoding and clean up imports

* Order metadata values by keys to hash them

* Import bitcoin functions

* Use list comprehensions instead of filter

Quoting python docs:

"map() and filter() return iterators. If you really need a list and the
input sequences are all of equal length, a quick fix is to wrap map() in
list(), e.g. list(map(...)), but a better fix is often to use a list
comprehension (especially when the original code uses lambda), or
rewriting the code so it doesn’t need a list at all. [...]"

ref: https://docs.python.org/3/whatsnew/3.0.html#views-and-iterators-instead-of-lists